@@ -1,31 +1,24 @@
 class CommentsController < ApplicationController
-  before_filter :get_parent
-
-  def new
-    @comment = @parent.comments.build
-  end
+  before_action :get_parent
 
   def create
-    @comment = @parent.comments.build(comment_params)
+    @comment = @parent.comments.new(comment_params)
+    @comment.user = current_user
     if @comment.save
-      redirect_to :back
+      redirect_back fallback_location: root_path, flash: { success: 'Успішно' }
     else
-      render :new
-      redirect_to :back
+      redirect_back fallback_location: root_path, flash: { error: 'Помилка' }
     end
   end
 
   protected
 
   def get_parent
-    @parent = Show.find(params[:show_id]) if params[:show_id]
-    @parent = Episode.find(params[:episode_id]) if params[:episode_id]
-    @parent = Comment.find(params[:comment_id]) if params[:comment_id]
-    redirect_to root_path unless defined?(@parent)
+    klass = comment_params[:commentable_type].constantize
+    @parent = klass.find(comment_params[:commentable_id])
   end
 
   def comment_params
-    params.require(:comment).permit(:body, :commentable_type, :commentable_id, :user_id)
+    params.require(:comment).permit(:body, :commentable_type, :commentable_id)
   end
-
 end
