@@ -3,12 +3,11 @@ class RaterController < ApplicationController
   def create
     if user_signed_in?
       obj = params[:klass].classify.constantize.find(params[:id])
-      obj.rate params[:score].to_i, current_user, params[:dimension]
+      obj.rate params[:score].to_f, current_user, params[:dimension]
 
-      render :json => true
+      render json: true
     else
-      render :json => false
+      render json: false
     end
   end
-
 end
