@@ -1,19 +1,25 @@
 $(document).ready(function () {
-    $('#form').submit(function(){
-        $.ajax({
-            type: "get",
-            url: "shows",
-            data: {
-                sort: $("#name").val(),
-                genre: $("#genre").val(),
-                page: $("#page").val(),
-                search: $("#search").val()
-            },
-            success: function(data){
-                data = $(data);
-                $("#content").html($('#shows', data).html());
-            }
-        });
-        return false;
+
+  var filter = debounce(function () {
+    var params = $('#form').serialize();
+    history.pushState(null, '', '?' + params);
+    $.get("/shows/list", params, function (data) {
+      $('#content').html(data);
     });
+  }, 250);
+
+
+  $('#form input[type=text]').keyup(filter);
+  $('#form select').add('#form input[type=checkbox]').change(filter);
+
+  function debounce(a, b, c) {
+    var d;
+    return function () {
+      var e = this, f = arguments;
+      clearTimeout(d), d = setTimeout(function () {
+        d = null, c || a.apply(e, f)
+      }, b), c && !d && a.apply(e, f)
+    }
+  }
+
 });
