@@ -1,1 +1,1 @@
-Rails.application.config.i18n.default_locale = :en
+Rails.application.config.i18n.default_locale = :uk
